backend/pyproject.toml
[project]
name = "alash-chatbot-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "python-jose>=3.3.0",
    "python-multipart>=0.0.6",
    "langchain>=0.3.0",
    "langchain-community>=0.0.10",
    "langchain-openai>=0.3.3",
    "langchain-chroma>=0.0.5",
    "chromadb>=0.6.3",
    "langchain-qdrant>=0.2.0",
    "chroma-hnswlib>=0.7.3",
    "bcrypt==4.0.1",
    "sqlalchemy>=2.0.23",
    "alembic>=1.12.1",
    "mysql-connector-python>=8.0.33",
    "minio>=7.2.0",
    "python-docx>=0.8.11",
    "pypdf>=3.0.0",
    "unstructured>=0.10.30",
    "unstructured[md]>=0.10.0",
    "markdown>=3.0.0",
    "openai>=1.30.0",
    "email-validator",
    "dashscope>=1.13.6",
    "langchain-deepseek==0.1.1",
    "langgraph>=0.2.0",
    "docx2txt==0.8",
    "openpyxl>=3.1.0",
    "pymupdf>=1.24.0",
]

[tool.uv]
package = false

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I"]

[dependency-groups]
dev = [
    "ruff>=0.15.8",
]
